requirements.txt
streamlit==1.29.0
numpy>=1.24.0
opencv-python>=4.8.0
pandas>=2.0.0
plotly>=5.15.0
seaborn>=0.12.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
Pillow>=10.0.0
python-dateutil>=2.8.0